move: ErrorMessage 컴포넌트 design폴더 -> common 폴더
