+++ src/components/common/Input/Input.tsx
@@ -1,7 +1,7 @@
 import type { Size } from '@/types/temp';
 import type { ComponentPropsWithRef, ForwardedRef, ReactElement } from 'react';
 import React, { forwardRef } from 'react';
-import ErrorMessage from '../../design/ErrorMessage/ErrorMessage';
+import ErrorMessage from '../ErrorMessage/ErrorMessage';
 import Label from '../Label/Label';
 import {
   getInputStyling,
